% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labs.R
\name{check_labs}
\alias{check_labs}
\title{Check for Missing Value Labels}
\usage{
check_labs(x)
}
\arguments{
\item{x}{Data set}
}
\value{
Vector of variables containing unlabelled values
}
\description{
Check for Missing Value Labels
}
