% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labs.R
\name{make_labs}
\alias{make_labs}
\title{Generate Variable and Value Labels}
\usage{
make_labs(
  x,
  out = "",
  max_labels = 30,
  max_cases = 30,
  fill = F,
  fill.limits = c(-1, 100)
)
}
\arguments{
\item{x}{Data set}

\item{out}{File Path for R Script}

\item{max_labels}{Maximum number of distinct value labels to be generated for a single variable}

\item{max_cases}{Maximum number of distinct values within a variable for said variable to be considered}

\item{fill}{Should the function create labels for values that are between values from the data set but are not themselves found}

\item{fill.limits}{Boundaries for fill}
}
\value{
R Script
}
\description{
Generate Variable and Value Labels
}
